Reimplement printing and formatting
Use the traverse_subtags() function to its full potential when
printing in different formats. In particular, scope can now be
indicated with clever use of the pre_action and post_action
parameters.

Additionally, the formatters global variable is now a dict of tuples,
with this format:

formatters = {
	"format-name": (init_action, pre_action, post_action, done_action)
}

The items of the tuple get called before, during (in preorder mode),
during (in postorder mode), and after traversing the tag tree while
printing. If that makes sense. See print_subtags() if it doesn't.